refactor(base-gen): extract ShouldRunPrecheckMixin (Tier-1b Task 4)
Move the pre-extraction should_run gate and its template hooks out of the
BaseGenerationService monolith into base_generation/_should_run.py:
_should_run_before_extraction, _build_should_run_prompt,
_collect_scoped_interactions_for_precheck, _get_precheck_interaction_query_kwargs,
_resolve_should_run_model. Bodies moved verbatim; annotation-only self stubs on
the mixin (incl. the _last_precheck_sessions Precheck->Billing seam).

ShouldRunPrecheckMixin is composed ahead of ConfigFilterMixin/StatusChangeMixin/ABC
so every method stays MRO-resolvable on BaseGenerationService: the enterprise
class-attribute string-path patch of _resolve_should_run_model still setattr-shadows
via MRO, and subclass overrides of the hooks still win. The 8 abstract methods stay
physically on the base ABC (__abstractmethods__ frozenset unchanged).

The cheap pre-filter helper _cheap_should_run_reject stays a module function in
base_generation_service (tests pin it to that module path); the gate imports it
function-locally to avoid a circular import while keeping the string-path patch
effective.
